Function that runs one GA metaepoch. Wrapper function for GA::ga.

Function that runs one GA metaepoch. Wrapper function for GA::ga.

Usage

ga_metaepoch(config_ga)

Arguments

config_ga

list of GA::ga params

Value

list with named fields: solution, population, value. A solution is a value of the decision variable giving the best fitness. A population is a matrix representing final population. Value is the value of a fitness function for the solution.

Examples

tree_height <- 3
empty_config_ga <- lapply(1:tree_height, function(x) {
  list()
})
ga_metaepoch(empty_config_ga)
#> function (fitness, suggestions, lower, upper, tree_level, minimize) 
#> {
#>     ga_fitness <- ifelse(minimize, function(x) {
#>         -1 * fitness(x)
#>     }, fitness)
#>     config <- config_ga[[tree_level]]
#>     legal_passed_param_names <- Filter(function(name) {
#>         name %in% methods::formalArgs(GA::ga)
#>     }, names(config))
#>     params <- list(maxiter = 5, popSize = nrow(suggestions))
#>     for (param_name in legal_passed_param_names) {
#>         params[param_name] <- config[param_name]
#>     }
#>     params$fitness <- ga_fitness
#>     params$lower <- lower
#>     params$upper <- upper
#>     params$suggestions <- suggestions
#>     params$type <- "real-valued"
#>     params$monitor <- FALSE
#>     GA <- do.call(GA::ga, params)
#>     value <- ifelse(minimize, GA@fitnessValue * -1, GA@fitnessValue)
#>     list(solution = c(GA@solution[1, ]), population = GA@population, 
#>         value = value)
#> }
#> <bytecode: 0x5568b8141f18>
#> <environment: 0x5568b8145bd0>