Print method for class "hms".

Print method for class "hms".

Usage

# S4 method for hms
print(x, ...)

Arguments

x

hms s4 object

...

other print arguments

Value

It does not return anything. The obvious side effect is output to the terminal.

Examples

f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
#> A list of standard deviations (sigma) or a function to create population should be provided.
#> Metaepoch: 0 Best fitness: 4.90764 
#> Metaepoch: 1 Best fitness: 4.983772 
#> Metaepoch: 2 Best fitness: 4.983772 
#> Metaepoch: 3 Best fitness: 4.983772 
#> Metaepoch: 4 Best fitness: 4.983772 
#> Metaepoch: 5 Best fitness: 4.983772 
#> Metaepoch: 6 Best fitness: 4.983772 
#> Metaepoch: 7 Best fitness: 4.983772 
#> Metaepoch: 8 Best fitness: 4.983772 
#> Metaepoch: 9 Best fitness: 4.983772 
print(result)
#> Formal class 'hms' [package "hmsr"] with 11 slots
#>   ..@ root_id                        : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   ..@ metaepoch_snapshots            :List of 10
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 2
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 1.45 2.66 4.22 -3.71 4.58 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 254
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.6 4.91 4.77 4.62 4.88 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:30] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.91
#>   .. .. .. ..@ best_solution      : Named num 4.91
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0111
#>   .. .. .. ..@ fitness_evaluations: num 254
#>   .. .. .. ..@ blocked_sprouts    : list()
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 3.34 3.8 3.09 3.56 3.86 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 508
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.98 4.92 4.85 4.92 4.93 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 128
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "74bb1e5b-914c-46bc-b69f-be27d6ac7611"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.95 4.96 4.9 4.83 4.98 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:15] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0214
#>   .. .. .. ..@ fitness_evaluations: num 636
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.18 3.97 4.24 3.82 4.77 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 766
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.95 4.61 4.71 4.92 4.89 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 256
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "74bb1e5b-914c-46bc-b69f-be27d6ac7611"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 -3.39 4.98 4.98 4.97 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 67
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0454
#>   .. .. .. ..@ fitness_evaluations: num 1089
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.236 3.943 4.353 3.86 0.725 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1032
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.94 4.93 4.98 4.94 4.96 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 376
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "74bb1e5b-914c-46bc-b69f-be27d6ac7611"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.22 4.98 4.98 4.46 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 129
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0226
#>   .. .. .. ..@ fitness_evaluations: num 1537
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.36 4.25 4.64 4.49 3.18 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1285
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.94 4.98 4.98 4.95 4.94 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 506
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "74bb1e5b-914c-46bc-b69f-be27d6ac7611"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.98 4.98 4.98 4.98 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 193
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0269
#>   .. .. .. ..@ fitness_evaluations: num 1984
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.39 4.58 -3.61 4.59 -4.75 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1535
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] -0.446 4.818 4.947 4.75 2.054 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 633
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "74bb1e5b-914c-46bc-b69f-be27d6ac7611"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.98 4.98 4.98 4.98 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 258
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0212
#>   .. .. .. ..@ fitness_evaluations: num 2426
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.36 4.43 4.56 4.74 4.61 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1796
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.98 4.98 4.43 4.94 3.24 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 759
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "74bb1e5b-914c-46bc-b69f-be27d6ac7611"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.98 4.98 4.98 4.98 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 325
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0209
#>   .. .. .. ..@ fitness_evaluations: num 2880
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.505 -0.262 4.458 4.607 4.378 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2050
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.66 4.46 -1.79 4.46 4.71 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 882
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "74bb1e5b-914c-46bc-b69f-be27d6ac7611"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.74 4.91 -3.21 -4.25 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 386
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0215
#>   .. .. .. ..@ fitness_evaluations: num 3318
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.66 4.46 -1.79 4.46 4.71 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 882
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.62 -2.7 4.61 4.6 4.64 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2314
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "74bb1e5b-914c-46bc-b69f-be27d6ac7611"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.95 4.94 4.57 4.94 4.93 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 451
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0144
#>   .. .. .. ..@ fitness_evaluations: num 3647
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.66 4.46 -1.79 4.46 4.71 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 882
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "74bb1e5b-914c-46bc-b69f-be27d6ac7611"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.95 4.94 4.57 4.94 4.93 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "f6c0e796-d378-41f7-a4dc-b5127d60efc2"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 451
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "c406ac0f-9c41-4837-82c6-0139a7db78d1"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.527 -0.433 4.744 4.735 4.178 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.91
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.91
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. .. ..$ : num 4.91
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2572
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.00912
#>   .. .. .. ..@ fitness_evaluations: num 3905
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.91
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   ..@ best_fitness                   : num 4.98
#>   ..@ best_solution                  : Named num 4.98
#>   .. ..- attr(*, "names")= chr "x1"
#>   ..@ total_time_in_seconds          : num 0.215
#>   ..@ total_metaepoch_time_in_seconds: num 0.201
#>   ..@ metaepochs_count               : num 10
#>   ..@ deme_population_sizes          : num [1:3] 60 30 15
#>   ..@ lower                          : num -5
#>   ..@ upper                          : num 5
#>   ..@ call                           : language hms(fitness = f, lower = -5, upper = 5)