Print method for class "hms".

Print method for class "hms".

Usage

# S4 method for hms
print(x, ...)

Arguments

x

hms s4 object

...

other print arguments

Value

It does not return anything. The obvious side effect is output to the terminal.

Examples

f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
#> A list of standard deviations (sigma) or a function to create population should be provided.
#> Metaepoch: 0 Best fitness: 4.882331 
#> Metaepoch: 1 Best fitness: 4.974049 
#> Metaepoch: 2 Best fitness: 4.992627 
#> Metaepoch: 3 Best fitness: 4.992627 
#> Metaepoch: 4 Best fitness: 4.992627 
#> Metaepoch: 5 Best fitness: 4.992627 
#> Metaepoch: 6 Best fitness: 4.992627 
#> Metaepoch: 7 Best fitness: 4.992627 
#> Metaepoch: 8 Best fitness: 4.992627 
#> Metaepoch: 9 Best fitness: 4.992627 
print(result)
#> Formal class 'hms' [package "hmsr"] with 11 slots
#>   ..@ root_id                        : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   ..@ metaepoch_snapshots            :List of 10
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 2
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.59 2.46 0.81 4.44 2.88 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.88
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 265
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.58 4.78 4.74 4.78 4.47 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:30] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.88
#>   .. .. .. ..@ best_solution      : Named num 4.88
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0139
#>   .. .. .. ..@ fitness_evaluations: num 265
#>   .. .. .. ..@ blocked_sprouts    : list()
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.27 4.41 4.35 4.23 4.52 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.93
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.93
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 514
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.9 4.67 4.73 4.97 4.8 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.97
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 125
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "37d9368e-19f6-4718-bfe9-5a78bacb868a"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.88 4.86 4.96 4.96 4.83 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:15] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.97
#>   .. .. .. ..@ best_solution      : Named num 4.97
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.022
#>   .. .. .. ..@ fitness_evaluations: num 639
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.78 -4.67 4.78 4.78 4.51 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.93
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.93
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 773
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.97 4.93 -1.27 4.82 4.87 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.97
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 256
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "37d9368e-19f6-4718-bfe9-5a78bacb868a"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] -1.12 4.96 4.99 1.97 4.97 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 66
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0296
#>   .. .. .. ..@ fitness_evaluations: num 1095
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.32 4.55 -2.47 4.74 4.34 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.93
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.93
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1042
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.8 4.91 4.58 4.87 4.94 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.97
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 386
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "37d9368e-19f6-4718-bfe9-5a78bacb868a"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 3.92 4.97 4.99 4.42 4.97 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 124
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0281
#>   .. .. .. ..@ fitness_evaluations: num 1552
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.93 4.62 4.6 4.93 4.46 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.93
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.93
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1295
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 2.22 4.97 4.92 4.94 4.97 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.97
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 515
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "37d9368e-19f6-4718-bfe9-5a78bacb868a"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.91 4.68 4.72 4.58 4.61 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 188
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.035
#>   .. .. .. ..@ fitness_evaluations: num 1998
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.49 4.58 4.53 4.65 4.55 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.93
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.93
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1544
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.95 4.79 4.91 4.97 4.34 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.97
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 634
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "37d9368e-19f6-4718-bfe9-5a78bacb868a"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.85 4.9 4.85 4.74 4.85 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 248
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0252
#>   .. .. .. ..@ fitness_evaluations: num 2426
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.75 4.85 4.39 4.93 4.93 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.93
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.93
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1794
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.94 4.89 4.96 4.92 4.94 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.97
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 754
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "37d9368e-19f6-4718-bfe9-5a78bacb868a"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.99 4.94 4.97 4.96 4.84 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 305
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0279
#>   .. .. .. ..@ fitness_evaluations: num 2853
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.91 4.82 4.63 4.82 -2.78 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.93
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.93
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2052
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.97 -1.32 4.97 4.86 4.95 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.97
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 870
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "37d9368e-19f6-4718-bfe9-5a78bacb868a"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.94 2.01 4.96 4.97 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 367
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0269
#>   .. .. .. ..@ fitness_evaluations: num 3289
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.97 -1.32 4.97 4.86 4.95 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.97
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 870
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.6 4.46 4.69 4.62 4.79 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.93
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.93
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2313
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "37d9368e-19f6-4718-bfe9-5a78bacb868a"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.99 4.97 4.98 4.98 4.97 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 418
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0181
#>   .. .. .. ..@ fitness_evaluations: num 3601
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.97 -1.32 4.97 4.86 4.95 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.97
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.97
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. .. ..$ : num 4.97
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.88
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 870
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "37d9368e-19f6-4718-bfe9-5a78bacb868a"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.99 4.97 4.98 4.98 4.97 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.97
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "9a577f75-db3c-4d06-9280-a0f74a445715"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 418
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "250788f9-3424-4ea2-ad03-f19737713897"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.705 4.867 4.606 3.309 0.866 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.93
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.93
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : Named num 4.88
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : num 4.88
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. .. ..$ : num 4.93
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2575
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0403
#>   .. .. .. ..@ fitness_evaluations: num 3863
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.93
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   ..@ best_fitness                   : num 4.99
#>   ..@ best_solution                  : Named num 4.99
#>   .. ..- attr(*, "names")= chr "x1"
#>   ..@ total_time_in_seconds          : num 0.268
#>   ..@ total_metaepoch_time_in_seconds: num 0.248
#>   ..@ metaepochs_count               : num 10
#>   ..@ deme_population_sizes          : num [1:3] 60 30 15
#>   ..@ lower                          : num -5
#>   ..@ upper                          : num 5
#>   ..@ call                           : language hms(fitness = f, lower = -5, upper = 5)