Print method for class "hms".

Print method for class "hms".

Usage

# S4 method for hms
print(x, ...)

Arguments

x

hms s4 object

...

other print arguments

Value

It does not return anything. The obvious side effect is output to the terminal.

Examples

f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
#> A list of standard deviations (sigma) or a function to create population should be provided.
#> Metaepoch: 0 Best fitness: 4.948773 
#> Metaepoch: 1 Best fitness: 4.989756 
#> Metaepoch: 2 Best fitness: 4.993636 
#> Metaepoch: 3 Best fitness: 4.993636 
#> Metaepoch: 4 Best fitness: 4.993636 
#> Metaepoch: 5 Best fitness: 4.993636 
#> Metaepoch: 6 Best fitness: 4.993636 
#> Metaepoch: 7 Best fitness: 4.993636 
#> Metaepoch: 8 Best fitness: 4.993636 
#> Metaepoch: 9 Best fitness: 4.993636 
print(result)
#> Formal class 'hms' [package "hmsr"] with 11 slots
#>   ..@ root_id                        : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   ..@ metaepoch_snapshots            :List of 10
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 2
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 3.39 4.56 2.73 3.99 4.13 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 246
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.98 4.99 4.86 4.65 4.96 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:30] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.95
#>   .. .. .. ..@ best_solution      : Named num 4.95
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.00784
#>   .. .. .. ..@ fitness_evaluations: num 246
#>   .. .. .. ..@ blocked_sprouts    : list()
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.51 4.55 4.09 3.73 4.86 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 499
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.77 4.95 4.37 4.92 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 133
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9c8a73a2-a59b-4ec0-b853-6caac520372f"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.87 4.89 4.99 4.87 4.96 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:15] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0271
#>   .. .. .. ..@ fitness_evaluations: num 632
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.51 4.54 4.01 4.53 4.71 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 760
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.95 4.99 4.96 4.98 4.96 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 265
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9c8a73a2-a59b-4ec0-b853-6caac520372f"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.97 4.97 4.94 4.95 4.97 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 59
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0165
#>   .. .. .. ..@ fitness_evaluations: num 1084
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.64 3.4 4.93 -4.63 1.04 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1018
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 3.96 -0.24 4.35 3.32 4.45 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 391
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9c8a73a2-a59b-4ec0-b853-6caac520372f"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.99 4.99 4.99 4.99 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 118
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0166
#>   .. .. .. ..@ fitness_evaluations: num 1527
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.65 4.79 4.94 4.83 -4.62 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1270
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.99 4.99 4.29 4.56 1.17 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 533
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9c8a73a2-a59b-4ec0-b853-6caac520372f"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.99 4.91 4.98 4.99 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 187
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0166
#>   .. .. .. ..@ fitness_evaluations: num 1990
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.94 1.97 4.75 4.88 4.9 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1537
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.83 4.69 4.77 4.76 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 665
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9c8a73a2-a59b-4ec0-b853-6caac520372f"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.99 4.99 4.98 4.98 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 246
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0167
#>   .. .. .. ..@ fitness_evaluations: num 2448
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.82 4.94 4.69 -3.06 4.85 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1801
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.84 4.74 4.87 4.98 4.83 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 790
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9c8a73a2-a59b-4ec0-b853-6caac520372f"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.99 4.99 4.99 4.99 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 308
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0165
#>   .. .. .. ..@ fitness_evaluations: num 2899
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.78 4.83 4.77 4.43 4.79 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2061
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 0.972 4.845 4.988 4.921 4.882 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 915
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9c8a73a2-a59b-4ec0-b853-6caac520372f"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.993 -0.697 4.991 4.991 4.992 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 371
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0165
#>   .. .. .. ..@ fitness_evaluations: num 3347
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 0.972 4.845 4.988 4.921 4.882 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 915
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.07 4.74 4.62 4.18 -3.52 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2319
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9c8a73a2-a59b-4ec0-b853-6caac520372f"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.9917 4.9914 -0.0383 4.9924 4.9919 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 432
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0114
#>   .. .. .. ..@ fitness_evaluations: num 3666
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 0.972 4.845 4.988 4.921 4.882 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 915
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "9c8a73a2-a59b-4ec0-b853-6caac520372f"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.9917 4.9914 -0.0383 4.9924 4.9919 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "66b30fed-cfd8-4562-8492-612a98d02998"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 432
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "d148b169-4c08-4232-a865-2559eed391b0"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.67 4.74 4.94 1.65 4.72 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.95
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.95
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. .. ..$ : num 4.95
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2577
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.00726
#>   .. .. .. ..@ fitness_evaluations: num 3924
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.95
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   ..@ best_fitness                   : num 4.99
#>   ..@ best_solution                  : Named num 4.99
#>   .. ..- attr(*, "names")= chr "x1"
#>   ..@ total_time_in_seconds          : num 0.153
#>   ..@ total_metaepoch_time_in_seconds: num 0.13
#>   ..@ metaepochs_count               : num 10
#>   ..@ deme_population_sizes          : num [1:3] 60 30 15
#>   ..@ lower                          : num -5
#>   ..@ upper                          : num 5
#>   ..@ call                           : language hms(fitness = f, lower = -5, upper = 5)