Print method for class "hms".

Print method for class "hms".

Usage

# S4 method for hms
print(x, ...)

Arguments

x

hms s4 object

...

other print arguments

Value

It does not return anything. The obvious side effect is output to the terminal.

Examples

f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
#> A list of standard deviations (sigma) or a function to create population should be provided.
#> Metaepoch: 0 Best fitness: 4.992133 
#> Metaepoch: 1 Best fitness: 4.9949 
#> Metaepoch: 2 Best fitness: 4.9949 
#> Metaepoch: 3 Best fitness: 4.9949 
#> Metaepoch: 4 Best fitness: 4.9949 
#> Metaepoch: 5 Best fitness: 4.9949 
#> Metaepoch: 6 Best fitness: 4.9949 
#> Metaepoch: 7 Best fitness: 4.9949 
#> Metaepoch: 8 Best fitness: 4.9949 
#> Metaepoch: 9 Best fitness: 4.9949 
print(result)
#> Formal class 'hms' [package "hmsr"] with 11 slots
#>   ..@ root_id                        : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   ..@ metaepoch_snapshots            :List of 10
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 2
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.77 3.71 4.92 -3.02 4.72 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 253
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.95 4.98 4.88 4.91 4.93 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:30] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.00892
#>   .. .. .. ..@ fitness_evaluations: num 253
#>   .. .. .. ..@ blocked_sprouts    : list()
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.86 4.25 4.36 4.51 4.98 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 516
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] -2.91 4.84 -4.92 4.93 1.29 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 122
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "26a575cf-2544-42d0-8a78-f2c29c01bf58"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.89 4.99 4.96 4.87 4.91 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:15] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0149
#>   .. .. .. ..@ fitness_evaluations: num 638
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.88 4.3 4.96 -4.08 3.74 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 783
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.19 4.76 4.99 4.99 4.64 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 247
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "26a575cf-2544-42d0-8a78-f2c29c01bf58"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.51 4.94 4.93 4.93 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 68
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0408
#>   .. .. .. ..@ fitness_evaluations: num 1098
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.27 4.55 4.3 4.41 4.77 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1052
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.82 4.38 4.21 4.11 4.81 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 370
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "26a575cf-2544-42d0-8a78-f2c29c01bf58"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.59 4.9 4.63 4.81 4.91 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 137
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.019
#>   .. .. .. ..@ fitness_evaluations: num 1559
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.37 4.68 4.67 4.69 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1307
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.93 4.76 4.85 2.13 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 500
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "26a575cf-2544-42d0-8a78-f2c29c01bf58"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.68 4.92 4.99 4.89 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 193
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0194
#>   .. .. .. ..@ fitness_evaluations: num 2000
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.77 4.99 4.78 2.71 4.69 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1559
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.15 4.94 4.86 4.99 4.83 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 631
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "26a575cf-2544-42d0-8a78-f2c29c01bf58"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.99 4.99 4.96 4.98 4.99 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 250
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0194
#>   .. .. .. ..@ fitness_evaluations: num 2440
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.808 4.83 -3.277 4.801 -0.407 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1825
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.94 4.92 4.94 4.91 4.93 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 760
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "26a575cf-2544-42d0-8a78-f2c29c01bf58"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.95 4.99 4.99 4.99 3.19 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 313
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0195
#>   .. .. .. ..@ fitness_evaluations: num 2898
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.98 4.33 4.68 4.88 4.73 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2077
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.973 4.97 4.85 0.328 4.993 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 886
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "26a575cf-2544-42d0-8a78-f2c29c01bf58"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.9 4.74 4.83 4.92 4.93 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 375
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0189
#>   .. .. .. ..@ fitness_evaluations: num 3338
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.973 4.97 4.85 0.328 4.993 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 886
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.99 4.82 4.99 1.49 4.61 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2335
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "26a575cf-2544-42d0-8a78-f2c29c01bf58"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.83 4.96 4.94 4.82 4.95 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 441
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0129
#>   .. .. .. ..@ fitness_evaluations: num 3662
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.973 4.97 4.85 0.328 4.993 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 886
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "26a575cf-2544-42d0-8a78-f2c29c01bf58"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.83 4.96 4.94 4.82 4.95 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "08281f03-f031-4dbe-9ee6-7128cfe52280"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 441
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "75add79c-f5d2-40b6-bf4e-1074309090e6"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.53 3.9 4.81 4.3 3.75 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.99
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.99
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. .. ..$ : num 4.99
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2591
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.99
#>   .. .. .. ..@ best_solution      : Named num 4.99
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.00822
#>   .. .. .. ..@ fitness_evaluations: num 3918
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.99
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   ..@ best_fitness                   : num 4.99
#>   ..@ best_solution                  : Named num 4.99
#>   .. ..- attr(*, "names")= chr "x1"
#>   ..@ total_time_in_seconds          : num 0.182
#>   ..@ total_metaepoch_time_in_seconds: num 0.17
#>   ..@ metaepochs_count               : num 10
#>   ..@ deme_population_sizes          : num [1:3] 60 30 15
#>   ..@ lower                          : num -5
#>   ..@ upper                          : num 5
#>   ..@ call                           : language hms(fitness = f, lower = -5, upper = 5)