Print method for class "hms".

Print method for class "hms".

Usage

# S4 method for hms
print(x, ...)

Arguments

x

hms s4 object

...

other print arguments

Value

It does not return anything. The obvious side effect is output to the terminal.

Examples

f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
#> A list of standard deviations (sigma) or a function to create population should be provided.
#> Metaepoch: 0 Best fitness: 4.722915 
#> Metaepoch: 1 Best fitness: 4.976773 
#> Metaepoch: 2 Best fitness: 4.976773 
#> Metaepoch: 3 Best fitness: 4.976773 
#> Metaepoch: 4 Best fitness: 4.976773 
#> Metaepoch: 5 Best fitness: 4.976773 
#> Metaepoch: 6 Best fitness: 4.976773 
#> Metaepoch: 7 Best fitness: 4.976773 
#> Metaepoch: 8 Best fitness: 4.976773 
#> Metaepoch: 9 Best fitness: 4.976773 
print(result)
#> Formal class 'hms' [package "hmsr"] with 11 slots
#>   ..@ root_id                        : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   ..@ metaepoch_snapshots            :List of 10
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 2
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.72 3.95 4.42 2.42 4.28 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.72
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 259
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.95 4.52 4.9 4.6 4.47 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:30] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.72
#>   .. .. .. ..@ best_solution      : Named num 4.72
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.00999
#>   .. .. .. ..@ fitness_evaluations: num 259
#>   .. .. .. ..@ blocked_sprouts    : list()
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 3.63 4.3 3.31 4.69 4.2 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.72
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 527
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.95 2.96 4.89 4.92 4.94 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 118
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "e5e624c5-5f0a-417e-8a81-3b304b090131"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.91 4.96 4.93 4.94 4.8 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:15] "" "" "" "" ...
#>   .. .. .. .. .. .. .. .. ..$ : chr "x1"
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num(0) 
#>   .. .. .. .. .. .. ..@ best_solution               : num(0) 
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch: list()
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 0
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0168
#>   .. .. .. ..@ fitness_evaluations: num 645
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.35 3.92 4.72 4.77 -1.32 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.77
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.77
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 784
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 3.135 4.836 4.932 0.729 4.964 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 249
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "e5e624c5-5f0a-417e-8a81-3b304b090131"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.98 4.95 4.95 2.33 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 1
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 52
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.021
#>   .. .. .. ..@ fitness_evaluations: num 1085
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.33 4.11 3.53 4.08 3.79 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.77
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.77
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1039
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.796 4.128 0.311 4.898 4.86 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 375
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "e5e624c5-5f0a-417e-8a81-3b304b090131"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.97 3.09 4.96 4.86 4.95 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 2
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 115
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0212
#>   .. .. .. ..@ fitness_evaluations: num 1529
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.03 4.41 4.01 4.41 4.25 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.89
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.89
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1288
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.92 4.87 4.96 4.46 4.98 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 506
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "e5e624c5-5f0a-417e-8a81-3b304b090131"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.95 4.96 4.96 4.95 4.98 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 3
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 171
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0208
#>   .. .. .. ..@ fitness_evaluations: num 1965
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.71 4.42 4.38 4.45 4.27 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.89
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.89
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1551
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.98 4.98 4.68 4.8 -3.91 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 642
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "e5e624c5-5f0a-417e-8a81-3b304b090131"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.97 4.96 4.97 4.97 4.97 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 4
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 232
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0211
#>   .. .. .. ..@ fitness_evaluations: num 2425
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.52 4.54 4.6 4.63 4.76 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.89
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.89
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 1812
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.55 4.65 4.72 4.74 4.21 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 772
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "e5e624c5-5f0a-417e-8a81-3b304b090131"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.97 4.96 4.97 4.97 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 5
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 291
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0209
#>   .. .. .. ..@ fitness_evaluations: num 2875
#>   .. .. .. ..@ blocked_sprouts    :List of 2
#>   .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.89 4.34 4.76 4.58 4.84 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.89
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.89
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 8
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2055
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.6 4.86 2.53 4.61 4.67 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 906
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "e5e624c5-5f0a-417e-8a81-3b304b090131"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.98 4.98 2.33 4.98 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 6
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 351
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0211
#>   .. .. .. ..@ fitness_evaluations: num 3312
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.6 4.86 2.53 4.61 4.67 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 906
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.73 4.67 4.89 4.39 4.66 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.89
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.89
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 9
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2308
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "e5e624c5-5f0a-417e-8a81-3b304b090131"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.98 4.98 4.98 4.98 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 416
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.0145
#>   .. .. .. ..@ fitness_evaluations: num 3630
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   .. ..$ :Formal class 'MetaepochSnapshot' [package "hmsr"] with 7 slots
#>   .. .. .. ..@ demes              :List of 3
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ population                  : num [1:30, 1] 4.6 4.86 2.53 4.61 4.67 ...
#>   .. .. .. .. .. .. ..@ level                       : num 2
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.72
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 906
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "e5e624c5-5f0a-417e-8a81-3b304b090131"
#>   .. .. .. .. .. .. ..@ population                  : num [1:15, 1] 4.98 4.98 4.98 4.98 4.98 ...
#>   .. .. .. .. .. .. ..@ level                       : num 3
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.98
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.98
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 7
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. .. ..$ : num 4.98
#>   .. .. .. .. .. .. ..@ sprout                      : Named num 4.98
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ parent_id                   : chr "dcd3616c-4c91-41d0-9053-6cb8a42bb1fd"
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 416
#>   .. .. .. .. .. .. ..@ is_active                   : logi FALSE
#>   .. .. .. .. ..$ :Formal class 'Deme' [package "hmsr"] with 11 slots
#>   .. .. .. .. .. .. ..@ id                          : chr "fec7054e-3af4-42f9-93b8-259e2e1cce37"
#>   .. .. .. .. .. .. ..@ population                  : num [1:60, 1] 4.7072 4.5567 4.6218 4.8896 -0.0264 ...
#>   .. .. .. .. .. .. ..@ level                       : num 1
#>   .. .. .. .. .. .. ..@ best_fitness                : num 4.89
#>   .. .. .. .. .. .. ..@ best_solution               : Named num 4.89
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_solutions_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.72
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.77
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. .. .. .. ..@ best_fitnesses_per_metaepoch:List of 10
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.72
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.77
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. .. ..$ : num 4.89
#>   .. .. .. .. .. .. ..@ sprout                      : NULL
#>   .. .. .. .. .. .. ..@ parent_id                   : NULL
#>   .. .. .. .. .. .. ..@ evaluations_count           : num 2569
#>   .. .. .. .. .. .. ..@ is_active                   : logi TRUE
#>   .. .. .. ..@ best_fitness       : num 4.98
#>   .. .. .. ..@ best_solution      : Named num 4.98
#>   .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ time_in_seconds    : num 0.00912
#>   .. .. .. ..@ fitness_evaluations: num 3891
#>   .. .. .. ..@ blocked_sprouts    :List of 1
#>   .. .. .. .. ..$ : Named num 4.89
#>   .. .. .. .. .. ..- attr(*, "names")= chr "x1"
#>   .. .. .. ..@ is_evolutionary    : logi TRUE
#>   ..@ best_fitness                   : num 4.98
#>   ..@ best_solution                  : Named num 4.98
#>   .. ..- attr(*, "names")= chr "x1"
#>   ..@ total_time_in_seconds          : num 0.177
#>   ..@ total_metaepoch_time_in_seconds: num 0.164
#>   ..@ metaepochs_count               : num 10
#>   ..@ deme_population_sizes          : num [1:3] 60 30 15
#>   ..@ lower                          : num -5
#>   ..@ upper                          : num 5
#>   ..@ call                           : language hms(fitness = f, lower = -5, upper = 5)